Revert "[INFRA] Changed the scope of some dependencies to compile from provided to enable running flink jobs from the IDE."
This reverts commit 95bda0b206089efa1cf3784d9140c34a8e7138b2.

This is to make the maven compilation faster.

Please note that this means that when you want to run
something that is in ...flink-jobs from the IDE, then you probably
need to change the scope of the Flink and the Scala library
dependencies to 'compile' from 'provided'.
(The symptom of not properly setting this is "NoClassDefFoundError"
when running from the IDE.)